Tree: Add templating engine for chat completions
The jinja analog in JS is nunjucks. There is a huggingface package
that adds minimal jinja support, but it doesn't have enough to make
parsing similar to TabbyAPI.

Nunjucks is marked as unsafe for user templates, but this is for
arbitrary user-provided templates while running. The flow in TabbyAPI/YALS
requires a user to create a template file and solely refer to that file's
identifier through the duration of app usage.

Therefore, there's no point in the app's lifecycle that allows unsafe
template injection and using nunjucks should be okay.

Signed-off-by: kingbri <8082010+bdashore3@users.noreply.github.com>