Bindings: Fix stop and banned strings
The previous trie check only checked for end of word once the buffer
is fully iterated upon. Therefore, partial matches were being missed
and not stopping/marking for rewind.

To fix this, add the end of word check after the node is reassigned
in the loop which sees if the word has completed.

For example, a stop string of "\n" will work with buffer of "hey\n".

Signed-off-by: kingbri <8082010+kingbri1@users.noreply.github.com>