Config + API: Add use_as_default key
The use_as_default key is the first layer in overrides when loading
a model. Let's say you set max_seq_len to 8192 in the config and always
want that to be the default fallback.

Well, this is where the user can specify if a param should be a fallback
on API request.

Signed-off-by: kingbri <8082010+kingbri1@users.noreply.github.com>