Bindings + Config: Add num_threads and chunk_size paramters
num_threads sets the amount of CPU threads to use for batch processing
and generation. Not applicable when model is fully GPU offloaded.

chunk_size, like exl2, sets the max amount of tokens to process at
a time. Tokens are usually divided into chunks to ease the amount
of VRAM.

Signed-off-by: kingbri <8082010+kingbri1@users.noreply.github.com>