Sampling: Add add_bos_token changes
Nowadays, add_bos_token is specified by the model. In chat completions,
use the template to determine whether a BOS token should be added, but
in text completions, users can either add or not add. However, if
null/undefined is provided, the model's preference is used.

Looking into it, this parameter should be removed entirely, but users
love to tinker.

Signed-off-by: kingbri <8082010+kingbri1@users.noreply.github.com>