Zod: Add cleanOptional method
YALS needs to accept both null and undefined values for a request,
but TypeScript treats the absence of a value as undefined, not null.

This causes an issue when using TS's optional syntax since the Zod
provided value is "value | null | undefined". Now, cleanOptional wraps
nullish and casts a null value to undefined which conforms to
TS's spec of "value | undefined".

Therefore, both null and undefined are accepted (unlike z.optional),
but only undefined is emitted.

Signed-off-by: kingbri <8082010+kingbri1@users.noreply.github.com>