Bindings: Defer cancellation cleanup to next loop iteration
Cancellation was causing resources such as rewinds to be freed which
caused illegal access errors. This was initially mitigated by migrating
the lock further back, but this also caused cancellation of hardware
to not occur due to holding the lock.

An easier approach is to defer cleanup until the next run loop iteration
which resets cancelled slots and marks them as ready to use.

Signed-off-by: kingbri <8082010+kingbri1@users.noreply.github.com>